Completed first version of algorithm and configured DI

+++ SPB_Path/SPB_Path.Domain/Places/Models/PlaceView.cs
@@ -9,6 +9,6 @@ namespace SPB_Path.Domain.Places.Models
     public class PlaceView
     {
         public string Name { get; set; }
-        public float RequiredTime { get; set; }
+        public double RequiredTime { get; set; }
     }
 }
